Add a setting for using locally installed Graphviz for generating diagrams.
The Google Charts service is only capable of rendering a limited range
of characters, so if you name your agents in an exotic language
generated diagrams are virtually useless being unable to make out.

With this change, Huginn embeds a locally generated SVG diagram instead
of relying on a third-party API if ENV['USE_GRAPHVIZ_DOT'] is set to a
command name/path of dot(1).

While at it:
- Use URI to compose a URL instead of hand-crafting it.
- Make the dot renderer a helper method.