ImapFolderAgent: Only keep a single UID value for each folder in memory.
Previously it used to keep a list of the UIDs of unread mails.  Now we
start to assume that UIDs in a folder identified by a UID VALIDITY value
are strictly ascending (monotonically increasing) as suggested by RFC
3501 and 4549 and just keep the highest UID seen in the last run.

This enhancement will help reduce the size of memory typically where
mails are left unread forever.