BasecampAgent use the remote time rather then ours
In the initial version of the BasecampAgent we used our own time to
query for new events. However that fails horrible if the clock of the
huginn server is out of sync the with clock of the api server.
Since we never can keep those clocks in sync we now rely on the
timestamp of the last event send to us.