Create database in `utf8` even if DATABASE_ENCODING is `utf8mb4`.
This is because we specify the `utf8mb4` charset in some of the columns
and expect others to be defaulted to `utf8`.

I'd like to make this a feature of `ar_mysql_column_charset`.  If you
want to set a table charset, you can always pass a keyword option
`options: "default charset=utf8mb4"` to `create_table`.

Complements #414.