Make event-indicator a link to the events page, with new events highlighted.
Currently it works as a reload button, but it gives you nothing but
dismissing the indicator unless you are on the events page, and in any
case you have no idea which events are newly created.

With this change, clicking the indicator takes you to the events page on
which newly created events are highlighted.