RssAgent: Use RSS::Parser directly
FeedNormalizer is no longer maintained, and its Atom support has flaws
in that it throws away what RSS::Parser returns and falls back to using
SimpleRSS which is not capable of handling XML entities, resulting in
getting ususable URLs such as ones including `&amp;`.

A breaking change is removal of the `clean` option which needs to be
reimplemented if it is important, while I personally think you should
use EventFormattingAgent to tweak feed contents if you need to.

This should address #889 and #955, among others.