Rescue ActiveRecord when database connection drops
Catch the exception ActiveRecord::StatementInvalid that appears when database connections are dropped unexpectedly when the database server is down, or killed by admin scripts on heroku free and basic plans. When the exception appears, it force the a new reconnection. Without this rescue, huginn was blocked (no new agent updates) when this situation appear.