Refactor agent forms into a model to clean up view
By extracting all of the logic from the view, we can make it more
declarative and easier to reason about. Form objects can help with that
by giving one place to look for all of the behavior of the form. It can
also help clean up the logic in controllers, but this PR doesn't get to
that much yet.

This also introduces some view helper classes called "Selectors" that
abstract the details of formatting data for Rails' `f.select` helper
that generates select boxes. Since there's heavy use of select2, there's
also a select2 version that helps to abstract the differences between
normal select boxes and select2 boxes.