Fix a problem in resolving the `url` key of a created event
The `url` parameter of handle_data() could hold a string or nil when
invoked from handle_event_data(), in which case resolving `url` in a
created event would fail with a type error.  Moreover, the logic did not
have any guard for URI errors.  This commit should fix them.

Fixes #1765.