docker: generate Rails secret at first run if not configured
After the container is created with "docker run", if the default token
("REPLACE_ME_NOW!") is still in .env.default, generate a random token
which will persist for the lifetime of the container (until the
container is rebuilt).

If the user sets APP_SECRET_TOKEN in the environment, this substitution
will be skipped and the env var will take precedence.