Close #1914 - remove 'become user' if deactivated user (#1938)
Ticket -- https://github.com/cantino/huginn/issues/1914
Testing -- Local huginn setup
About --
Switching as admin to a user is broken if the user is deactivated.

This code makes the 'Become User' buttons invisible if the user to
be impersonated is deactivated. The buttons used to be visible on:
* /admin/users
* /admin/users/XX/edit

Bonus: removes the 'Become User' button from the 'Edit Users' page
if we are the edited `current_user`.