Create Tumblr Likes Agent (#1923)
* Create Tumblr Likes Agent

* Don't interpolate options

* Clear ID memory after 50 entries

In order to prevent the ID array from getting to large, keep only the
last 50 entries. 50 is currently the maximum the API will return at a
time, so we should never exceed that limit.

* Request likes after the last recorded like

Instead of just returning blog likes, request only likes after the
timestamp of the last liked item we’ve processed.