Fix flaky spec
The `select_agent_type` helper is waiting for the options editor to be
initialized, which will never happen for Agents that are
FormConfigurable. The GrowlAgent was recently changed to
FormConfigurable which mean the spec should always have failed.
I have no idea how it passes on CI in some cases but this should fix the
'random' failures.