Fix HttpStatusAgent case insensitive header lookup
Ruby 2.5 added the `Hash#slice` method which allowed Rails 5.2 [to not
apply][1] it's ActiveSupport core extension. Because the Ruby
implementation is done in C it [does not call the `[]` method][2] which
is overridden by [`Faraday::Utils::Headers`][3] to achieve case
insensitive header look ups.

By calling the `Faraday::Utils::Headers#[]` directly we keep the case
insensitive lookup.

[1]: https://github.com/rails/rails/commit/a524c955bfcce118da0ecb21affe28e2a102a54b
[2]: https://github.com/ruby/ruby/blob/ae92a9e4ed7665130c816ae3f8a6927242d367ad/hash.c#L1357-L1375
[3]: https://github.com/lostisland/faraday/blob/4381c2d62da4f148085c0863a20f441625ad3791/lib/faraday/utils.rb#L51-L54