Do not seed when a user exists
Previously the seed logic ran when `SEED_EMAIL` was unset and the `admin`
users email address was changed, or if `SEED_EMAIL` did not match it
anymore.

We run `db:seed` on every docker container start to make it easy to get
started with Huginn. That could lead to situations in which the
container did not start if the admin users email address was changed (the
Seeder failed due to the uniqueness constraint on `username`).

Since we "seed" the database I think it is best to bail out of the
seeding logic as long as at least one user exists.

Fixes #2362