Set inbound_event when creating AgentLog entries
If an Agent is using `interpolate_with_each` to when can automatically
set the `inbound_event` for every `AgentLog` entry the Agent creates.

This is especially useful when using a `WebsiteAgent` to work on
different sites. Before this changes error log entries lacked to
context of the incoming Event to debug the failure.