Backfill CHANGES.md
In the past we mostly only included breaking changes or if the
contributor felt like grabbing a spot in the changelog. I think it is
easier to include every PR we merge.

I updated the [script used to generate it](https://gist.github.com/dsander/f9e9a0a4043415bc108e76157674e8de).