[docker] Fix mysql volume usage for new multi-process containers
The Ubuntu upgrade (or changes to MySQL after the update was initially
tested) broke the usage of mounting a directory as the MySQL data volume
for the multi-process image.

Why `--admin-auth-plugin=mysql_native_password` was needed (or I thought
it was needed) is beyond me now. Passing the `mysqld` process the
`--init-file` option which creates our database user seems to work in
all use cases.

 #2745
 #2757