Hopefully fix ruby segfault in docker with Rails 6
Something in the docker layer cache must be too old or incompatible with
previous layers. This change should force a full rebuild of the
container.

It is possible that we need to abandon the layer caching optimizations
on all CI systems to ensure we are actually validating changes based on
similar (enough) images.

 #2605