Fix liquid interpolation cache on ruby 2.7
I can not really explain why this broke. For some reason the array we
use as an index for the interpolation cache was not "hashed" properly
and created false hits inn the cache. Explicitly calling `hash` on the
array key also calls `hash` on elements and properly caches the data.