Create a passwd entry for the default user during docker build
docker/scripts/setup_env includes a piece of code that tries to do this, but it is executed in run time as the default user who does not have permissions for that.

Running useradd(8) in Dockerfile as root instead ensures a passwd entry for the default user, which allows commands like ssh to work in a ShellCommandAgent.

I made UID tunable via the --build-arg option while I'm here.