Push docker images to docker hub and ghcr
As discussed in #3234 we want to make the transition from hosting the
docker images on docker hub to GitHub container registry. This change
will push the images to both registries, after we have verified
everything works as expected we can embed a warning in images pushed to
docker hub which tells the users to switch.

[This is the github action run][1] on my fork that successfully created
the [images][2].

I didn't bother with integrating the two registries into a build matrix
because it is probably easier to just remove the docker hub jobs once we
are done with the transition.

[1]: https://github.com/dsander/huginn/actions/runs/4451875730
[2]: https://github.com/dsander?tab=packages&repo_name=huginn